secure foreach in threadpool
git-svn-id: http://manialive.googlecode.com/svn/trunk@224 282cbf28-0b21-c27b-f752-b18d30e2b6b0

+++ libraries/ManiaLive/Threading/ThreadPool.php
@@ -69,15 +69,17 @@ class ThreadPool extends \ManiaLib\Utils\Singleton implements \ManiaLive\Feature
 			// keep manialive clean!
 			// delete unused threading databases ...
 			$files = glob(APP_ROOT . '/data/threading_*.db'); 
-			foreach($files as $file)
+			if(is_array($files))
 			{
-				try
-				{
-					unlink($file);
-				}
-				catch (\Exception $e) {}
+			    foreach($files as $file)
+			    {
+				    try
+				    {
+					    unlink($file);
+				    }
+				    catch (\Exception $e) {}
+			    }
 			}
-			
 			$this->database = Tools::getDb();
 			
 			// setup database ...
